# feat: cache `Indent` and use `PadRight`