Add cache to `IncreaseIndent`and use `PadRight`. Saves 4MB 6% of memory.

- Could this be a `List<Indent>`? If I understand correctly csharpier won't change between tabs and space mode, so this should be okay right?
- Assumes that `TypesForTabs` is only used for tabs mode.
- `AddAlign` and `MakeIndentWithTypesForTabs` could be improved, I should run the benchmarks with tabs mode enabled to see if its worth it.

### Before
| Default_CodeFormatter | 234.9 ms | 11.21 ms | 33.05 ms | 7000.0000 | 3000.0000 | 1000.0000 |  66.17 MB |

### After
| Default_CodeFormatter | 223.6 ms | 9.96 ms | 29.36 ms | 212.7 ms | 7000.0000 | 3000.0000 | 1000.0000 |  62.18 MB |