@belav 2025-02-25
> Could this be a List<Indent>? If I understand correctly csharpier won't change between tabs and space mode, so this should be okay right?

Are you referring to the dictionary? I'm guessing the idea is using an index into the list where the index is the `indent.value`.

CSharpier doesn't switch between tabs and spaces in a file, but the `Align` doc would add some spaces after tabs in some cases. But I don't actually see any usages of it. 

I'm guessing it was used when there was some formatting that would mix a 2 space indent into the usual 4 space indents. In the case of tabs it would add 2 spaces after the tab indent level. I'm in favor of ripping out `Align`. There was pushback to the mixing of indent level and I don't forsee any new formatting that would use it.

I approved but let me know if you want to make more changes here. I can also take care of ripping out `Align` unless you'd want to deal with it at the same time. It may result in some conflicts.