@belav 2025-02-28
I removed align in this PR https://github.com/belav/csharpier/pull/1516